lambda/inference/requirements.txt
numpy<2.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
boto3>=1.26.0
